update api url

+++ client/src/features/user/userApiSlice.js
@@ -30,9 +30,7 @@ export const deletePermission = createAsyncThunk(
   'user/deletePermission',
   async (id) => {
     try {
-      const res = await api.delete(
-        `https://dashboard-oubj.onrender.com/api/v1/permission/${id}`
-      );
+      const res = await api.delete(`/permission/${id}`);
       return res.data;
     } catch (error) {
       throw new Error(error.response.data.message);
@@ -44,10 +42,7 @@ export const updatePermission = createAsyncThunk(
   'user/updatePermission',
   async ({ id, data }) => {
     try {
-      const res = await api.patch(
-        `https://dashboard-oubj.onrender.com/api/v1/permission/${id}`,
-        data
-      );
+      const res = await api.patch(`/permission/${id}`, data);
       return res.data;
     } catch (error) {
       throw new Error(error.response.data.message);
@@ -75,9 +70,7 @@ export const addRole = createAsyncThunk('user/addRole', async (data) => {
 
 export const deleteRole = createAsyncThunk('user/deleteRole', async (id) => {
   try {
-    const res = await api.delete(
-      `https://dashboard-oubj.onrender.com/api/v1/role/${id}`
-    );
+    const res = await api.delete(`/role/${id}`);
     return res.data;
   } catch (error) {
     throw new Error(error.response.data.message);
@@ -88,10 +81,7 @@ export const updateRole = createAsyncThunk(
   'user/updateRole',
   async ({ id, data }) => {
     try {
-      const res = await api.patch(
-        `https://dashboard-oubj.onrender.com/api/v1/role/${id}`,
-        data
-      );
+      const res = await api.patch(`/role/${id}`, data);
       return res.data;
     } catch (error) {
       throw new Error(error.response.data.message);
@@ -118,9 +108,7 @@ export const addUser = createAsyncThunk('user/addUser', async (data) => {
 
 export const deleteUser = createAsyncThunk('user/deleteUser', async (id) => {
   try {
-    const res = await api.delete(
-      `https://dashboard-oubj.onrender.com/api/v1/user/${id}`
-    );
+    const res = await api.delete(`/user/${id}`);
     return res.data;
   } catch (error) {
     throw new Error(error.response.data.message);
@@ -131,10 +119,7 @@ export const updateUser = createAsyncThunk(
   'user/updateUser',
   async ({ id, data }) => {
     try {
-      const res = await api.patch(
-        `https://dashboard-oubj.onrender.com/api/v1/user/${id}`,
-        data
-      );
+      const res = await api.patch(`/user/${id}`, data);
       return res.data;
     } catch (error) {
       throw new Error(error.response.data.message);
